Move pass-and-play related API utilities to dedicated utility module
